+++ src/ui/imgui_custom.cpp
@@ -38,7 +38,7 @@ namespace ImGui
 
     IMGUI_API bool DragDouble(const char* label, double* v, double v_speed, double v_min, double v_max, const char* format, ImGuiSliderFlags flags)
     {
-        return DragScalar(label, ImGuiDataType_Double, v, v_speed, &v_min, &v_max, format, flags);
+        return DragScalar(label, ImGuiDataType_Double, v, (float)v_speed, &v_min, &v_max, format, flags);
     }
 
     IMGUI_API bool SliderDouble2(const char* label, double v[2], double v_min, double v_max, const char* format, ImGuiSliderFlags flags)
@@ -48,7 +48,7 @@ namespace ImGui
 
     IMGUI_API bool DragDouble2(const char* label, double v[2], double v_speed, double v_min, double v_max, const char* format, ImGuiSliderFlags flags)
     {
-        return DragScalarN(label, ImGuiDataType_Double, v, 2, v_speed, &v_min, &v_max, format, flags);
+        return DragScalarN(label, ImGuiDataType_Double, v, 2, (float)v_speed, &v_min, &v_max, format, flags);
     }
 
     bool SymbolButton(const char* id, const char* symbol, const ToolbarButtonState& state, ImVec2 size)
